moved interfaces of coordinatesForm into interfaces/ directory
